#355: Model struct별 파일 분리
