--- Todoary-iOS/Todoary/Network/Entity/Auth/AccountModel.swift
@@ -1,41 +1,0 @@
-//
-//  AccountModel.swift
-//  Todoary
-//
-//  Created by 박지윤 on 2023/01/10.
-//
-
-import Foundation
-
-struct SignUpRequestModel: Codable{
-    var email: String?
-    var name: String?
-    var nickname: String?
-    var password: String?
-    var isTermsEnable: Bool?
-}
-
-struct AppleSignUpRequestModel: Codable{
-    let code: String //authorizationCode
-    let idToken: String
-    let name: String
-    let email: String
-    let isTermsEnable: Bool
-    let userIdentifier: String
-}
-
-struct AppleSignUpResultModel: Codable{
-    let isNewUser: Bool
-    let name: String
-    let email: String
-    let provider: String
-    let providerId: String
-    let token: Token
-    let appleRefreshToken: String
-}
-
-
-struct DeleteAppleAccountRequestModel: Codable{
-    let email: String
-    let code: String
-}
