+++ Todoary-iOS/Todoary/Network/Entity/Auth/AppleSignUpResultModel.swift
@@ -1,0 +1,18 @@
+//
+//  AppleSignUpResultModel.swift
+//  Todoary
+//
+//  Created by 박지윤 on 2023/01/17.
+//
+
+import Foundation
+
+struct AppleSignUpResultModel: Codable{
+    let isNewUser: Bool
+    let name: String
+    let email: String
+    let provider: String
+    let providerId: String
+    let token: Token
+    let appleRefreshToken: String
+}
