+++ Todoary-iOS/Todoary/Network/Entity/Auth/SignUpRequestModel.swift
@@ -1,0 +1,16 @@
+//
+//  SignUpRequestModel.swift
+//  Todoary
+//
+//  Created by 박지윤 on 2023/01/17.
+//
+
+import Foundation
+
+struct SignUpRequestModel: Codable{
+    var email: String?
+    var name: String?
+    var nickname: String?
+    var password: String?
+    var isTermsEnable: Bool?
+}
