+++ Todoary-iOS/Todoary/Network/Entity/Diary/DiaryRequestModel.swift
@@ -1,0 +1,13 @@
+//
+//  DiaryRequestModel.swift
+//  Todoary
+//
+//  Created by 박지윤 on 2023/01/17.
+//
+
+import Foundation
+
+struct DiaryRequestModel: Codable{
+    let title: String
+    let content: String
+}
