+++ Todoary-iOS/Todoary/Network/Entity/User/AlarmActiveStautsResultModel.swift
@@ -1,5 +1,5 @@
 //
-//  UserModel.swift
+//  AlarmActiveStautsResultModel.swift
 //  Todoary
 //
 //  Created by 박지윤 on 2023/01/10.
